Refactor QR vector layout and fix XML docs
